+++ frontend/src/components/settings/settings-drawer.vue
@@ -1,11 +1,12 @@
 <template>
   <v-navigation-drawer
-    v-model="isSettingsDrawerOpen"
-    class="settingsDrawer"
-    app
-    location="right"
-    touchless
     v-bind="$attrs"
+    v-model="isSettingsDrawerOpen"
+    disable-resize-watcher
+    disable-route-watcher
+    location="right"
+    temporary
+    touchless
   >
     <div class="settingsDrawerContainer">
       <div id="topBlock">
@@ -13,7 +14,7 @@
           <h3>{{ title }}</h3>
         </header>
         <component :is="panel" v-if="isCodexViewable" />
-        <v-divider />
+        <v-divider v-if="isCodexViewable" />
         <SettingsCommonPanel :admin-menu="adminMenu" />
         <v-divider />
       </div>
@@ -28,13 +29,14 @@
 </template>
 
 <script>
-import { mapGetters, mapWritableState } from "pinia";
+import { mapGetters, mapState, mapWritableState } from "pinia";
 
 import SettingsCommonPanel from "@/components/settings/panel.vue";
 import SettingsFooter from "@/components/settings/settings-footer.vue";
 import VersionFooter from "@/components/settings/version-footer.vue";
 import { useAuthStore } from "@/stores/auth";
 import { useCommonStore } from "@/stores/common";
+import { useReaderStore } from "@/stores/reader";
 
 export default {
   name: "SettingsDrawer",
@@ -60,6 +62,9 @@ export default {
   computed: {
     ...mapGetters(useAuthStore, ["isCodexViewable"]),
     ...mapWritableState(useCommonStore, ["isSettingsDrawerOpen"]),
+    ...mapState(useReaderStore, {
+      invisibleHack: (state) => state.bookChange === "next",
+    }),
   },
   mounted() {
     this.isSettingsDrawerOpen =