+++ NEWS.md
@@ -6,6 +6,14 @@ width: 128px;
 border-radius: 128px;
 " />
 
+## v1.0.1
+
+- Features
+  - Justify order-by field in browser cards.
+- Fixes
+  - Fixed next book change drawer opening settings drawer.
+  - Fixed zero padding on browser card issue numbers.
+
 ## v1.0.0
 
 ### Vue 3